+++ front-end/tailwind.config.js
@@ -3,6 +3,7 @@ module.exports = {
 	content: [
 		"./src/**/*.{html,js,jsx,tsx}",
 		"./node_modules/react-tailwindcss-datepicker/dist/index.esm.js",
+		"node_modules/preline/dist/*.js",
 	],
 	theme: {
 		extend: {